refactoring dashboard
